# openapi/community.json
{
  "components": {
    "schemas": {
      "AnthropicErrorDetail": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "AnthropicErrorDetail",
        "type": "object"
      },
      "AnthropicErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/AnthropicErrorDetail"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "error"
        ],
        "title": "AnthropicErrorResponse",
        "type": "object"
      },
      "AnthropicModelListView": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AnthropicModelRecordView"
            },
            "title": "Data",
            "type": "array"
          },
          "first_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Id"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "last_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Id"
          }
        },
        "required": [
          "data",
          "has_more",
          "first_id",
          "last_id"
        ],
        "title": "AnthropicModelListView",
        "type": "object"
      },
      "AnthropicModelRecordView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "display_name",
          "type"
        ],
        "title": "AnthropicModelRecordView",
        "type": "object"
      },
...
      "ChatRequest": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "title": "ChatRequest",
        "type": "object"
      },
...
      "CodexModelListView": {
        "properties": {
          "models": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "required": [
          "models"
        ],
        "title": "CodexModelListView",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "ModelListView": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ModelRecordView"
            },
            "title": "Data",
            "type": "array"
          },
          "object": {
            "title": "Object",
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "ModelListView",
        "type": "object"
      },
      "ModelRecordView": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object": {
            "title": "Object",
            "type": "string"
          },
          "owned_by": {
            "title": "Owned By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "object",
          "created",
          "owned_by"
        ],
        "title": "ModelRecordView",
        "type": "object"
      },
...
      "OpenAIErrorDetail": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "param": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "message",
          "type"
        ],
        "title": "OpenAIErrorDetail",
        "type": "object"
      },
      "OpenAIErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/OpenAIErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "OpenAIErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AnthropicAPIKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "ShimAPIKey": {
        "in": "header",
        "name": "x-shim-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "SHIM Trust Boundary Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "operationId": "chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Native Chat Completion JSON or server-sent events."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "408": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          },
          "529": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                }
              }
            },
            "description": "Sanitized OpenAI-compatible error."
          }
        },
        "security": [
          {
            "ShimAPIKey": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat Completions"
      }
    },
...
    "/v1/models": {
      "get": {
        "operationId": "list_models_v1_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "client_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Version"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          },
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "anthropic-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anthropic-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ModelListView"
                    },
                    {
                      "$ref": "#/components/schemas/CodexModelListView"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicModelListView"
                    }
                  ],
                  "title": "Response List Models V1 Models Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 400 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 401 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 403 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 404 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 422 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 429 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 503 List Models V1 Models Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          }
        },
        "security": [
          {
            "ShimAPIKey": []
          },
          {
            "HTTPBearer": []
          },
          {
            "AnthropicAPIKey": []
          }
        ],
        "summary": "List Models"
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "operationId": "retrieve_model_v1_models__model_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "anthropic-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Anthropic-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ModelRecordView"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicModelRecordView"
                    }
                  ],
                  "title": "Response Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 400 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 401 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 403 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 404 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 422 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 429 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AnthropicErrorResponse"
                    }
                  ],
                  "title": "Response 503 Retrieve Model V1 Models  Model Id  Get"
                }
              }
            },
            "description": "OpenAI- or Anthropic-compatible error selected by headers."
          }
        },
        "security": [
          {
            "ShimAPIKey": []
          },
          {
            "HTTPBearer": []
          },
          {
            "AnthropicAPIKey": []
          }
        ],
        "summary": "Retrieve Model"
      }
    },
...
  }
}